requirements.txt
flask==2.3.2
werkzeug==2.3.6
argon2-cffi==21.3.0
python-dotenv==1.0.0
filetype==1.2.0